Timer working, updating on screen
Location updates not yet working.

+++ app/src/main/java/edu/uwm/cs/fitrpg/FitnessActivitySegment.java
@@ -46,7 +46,7 @@ public class FitnessActivitySegment {
         return distance;
     }
 
-    public double getDuration() {
+    public long getDuration() {
         return stopTime - startTime;
     }
